Conflicts:
	django/contrib/admin/templatetags/admin_list.py
	django/forms/utils.py
	django/forms/widgets.py
	tests/utils_tests/test_html.py

+++ tests/expressions_regress/models.py
@@ -8,7 +8,7 @@ from django.db import models
 
 @python_2_unicode_compatible
 class Number(models.Model):
-    integer = models.IntegerField(db_column='the_integer')
+    integer = models.BigIntegerField(db_column='the_integer')
     float = models.FloatField(null=True, db_column='the_float')
 
     def __str__(self):
@@ -26,4 +26,3 @@ class Experiment(models.Model):
 
     def duration(self):
         return self.end - self.start
-
